requirements.txt
paho-mqtt==1.5.0
python-dotenv==0.10.5
